feat(website): optimize mobile examples and isolate showcase loading
