+++ website/src/examples.tsx
@@ -18,106 +18,84 @@ import type {Device} from '@luma.gl/core';
 import type {AnimationProps} from '@luma.gl/engine';
 
 import {makeHtmlCustomPanel} from '../../examples/example-panels';
-import {makeArrowExamplePanelHostHtml} from '../../examples/arrow/arrow-example-panels';
-import AnimationApp from '../../examples/api/animation/app';
-import BlendingApp from '../../examples/api/blending/app';
-import CubemapApp from '../../examples/api/cubemap/app';
-import ArrowDggsPolygonsApp from '../../examples/arrow/arrow-dggs-polygons/app';
-import ArrowColumnRendererApp from '../../examples/arrow/arrow-columns/app';
-import ArrowMeshGeometryApp from '../../examples/arrow/arrow-mesh-geometry/app';
-import ArrowGeoArrowApp from '../../examples/arrow/arrow-geoarrow/app';
-import ArrowLinesApp from '../../examples/arrow/arrow-lines/app';
-import ArrowFloat64PrecisionApp from '../../examples/arrow/arrow-float64-precision/app';
-import ArrowPointRendererApp from '../../examples/arrow/arrow-points/app';
-import ArrowFilteringApp from '../../examples/arrow/arrow-filtering/app';
-import ArrowPolygonRendererApp from '../../examples/arrow/arrow-polygons/app';
-import BloomApp from '../../examples/experimental/bloom/app';
-import HTMLUIPrismApp from '../../examples/experimental/html-ui-prism/app';
-import GPUFrustumCullingApp from '../../examples/experimental/gpu-frustum-culling/app';
-import GPUSceneGraphApp from '../../examples/experimental/gpu-scene-graph/app';
-import GPUTraceSceneApp from '../../examples/experimental/gpu-trace-scene/app';
-import GPUTraceViewerApp from '../../examples/experimental/gpu-trace-viewer/app';
-import GPUGraphExplorerApp from '../../examples/experimental/gpu-graph-explorer/app';
-import LuCIMVolumeLabApp from '../../examples/experimental/lucim-volume-lab/app';
-import {
-  initializeGPUSortExample,
-  type GPUSortExampleHandle
-} from '../../examples/experimental/gpu-sort/src/app';
-import {
-  initializeGPUDataAnalysisExample,
-  type GPUDataAnalysisExampleHandle
-} from '../../examples/experimental/gpu-data-analysis/src/app';
-import GPT2App from '../../examples/experimental/gpt-2/app';
-import VideoTextureApp from '../../examples/api/video-texture/app';
-import WebXRKaleidoscopeApp from '../../examples/experimental/webxr-kaleidoscope/app';
-import {
-  initializeGPGPUShowcase,
-  type GPGPUShowcaseHandle
-} from '../../examples/v10/gpgpu/src/app';
-import ArrowParticlesApp from '../../examples/arrow/arrow-particles/app';
-import MultiCanvasApp from '../../examples/api/multi-canvas/app';
-import Texture3DApp from '../../examples/api/texture-3d/app';
-import TextureSamplingApp from '../../examples/api/texture-sampling/app';
-import TextureTesterApp from '../../examples/api/texture-tester/app';
-import initializeExternalWebGLContext, {
-  ExternalWebGLContextHandle
-} from '../../examples/integrations/external-context/app';
-import HelloReactApp from '../../examples/integrations/hello-react/app';
-import {getErrorMessage, logError} from './react-luma/utils/error-utils';
-import DOFApp from '../../examples/showcase/dof/app';
-import AdvancedEffectsApp from '../../examples/experimental/advanced-effects/app';
-import DeferredRenderingApp from '../../examples/experimental/deferred-rendering/app';
-import FluidFoundryApp from '../../examples/experimental/fluid-foundry/app';
-import SpectralCausticsApp from '../../examples/experimental/spectral-caustics/app';
-import VolumetricFireForgeApp from '../../examples/experimental/volumetric-fire-forge/app';
-import VirtualGeometryCanyonApp from '../../examples/experimental/virtual-geometry-canyon/app';
-import ShadowMapApp from '../../examples/experimental/shadow-map/app';
-import ABufferApp from '../../examples/experimental/a-buffer/app';
-
-// import PerformanceApp from '../../examples/performance/stress-test/app';
-
-// import DOFApp from '../../examples/showcase/dof/app';
-// import GeospatialApp from '../../examples/showcase/geospatial/app';
-import GLTFApp from '../../examples/showcase/gltf/app';
-import GaussianSplatsApp from '../../examples/showcase/gaussian-splats/app';
+import type {GPUSortExampleHandle} from '../../examples/experimental/gpu-sort/src/app';
+import type {GPUDataAnalysisExampleHandle} from '../../examples/experimental/gpu-data-analysis/src/app';
+import type {GPGPUShowcaseHandle} from '../../examples/v10/gpgpu/src/app';
+import type {ExternalWebGLContextHandle} from '../../examples/integrations/external-context/app';
 import type {GaussianSplatSourceCatalogEntry} from '../../examples/showcase/gaussian-splats/local-loaders';
-import ArrowTemporalStarfieldApp from '../../examples/arrow/arrow-temporal-starfield/app';
-import ArrowTimeColumnsApp from '../../examples/arrow/arrow-time-columns/app';
-import ArrowText2DApp from '../../examples/arrow/arrow-text-2d/app';
-import InstancingApp from '../../examples/showcase/instancing/app';
-import LightstormMegacityApp from '../../examples/showcase/lightstorm-megacity/app';
-import LLMNetworkApp from '../../examples/showcase/llm-network/app';
-import VectorFieldLabApp from '../../examples/showcase/vector-field-lab/app';
-import QuantumStateStudioApp from '../../examples/showcase/quantum-state-studio/app';
-import TempestOceanApp from '../../examples/showcase/tempest-ocean/app';
-import RenderBundlesApp from '../../examples/api/render-bundles/app';
-import TextSpaceCrawlApp from '../../examples/experimental/text-space-crawl/app';
-import PersistenceApp from '../../examples/showcase/persistence/app';
-import PostprocessingApp from '../../examples/showcase/postprocessing/app';
-import AntialiasingApp from '../../examples/experimental/antialiasing/app';
-import GlobeApp from '../../examples/showcase/globe/app';
-import PacketSprayingApp from '../../examples/showcase/packet-spraying/app';
-// import WanderingApp from '../../examples/showcase/wandering/app';
-
-import HelloTriangleGeometryApp from '../../examples/tutorials/hello-triangle-geometry/app';
-import HelloTriangleApp from '../../examples/tutorials/hello-triangle/app';
-import HelloCubeApp from '../../examples/tutorials/hello-cube/app';
-import TwoCubesApp from '../../examples/tutorials/hello-two-cubes/app';
-import InstancedCubesApp from '../../examples/tutorials/hello-instanced-cubes/app';
-import HelloInstancingApp from '../../examples/tutorials/hello-instancing/app';
-import HelloGLTFApp from '../../examples/tutorials/hello-gltf/app';
-import LightingApp from '../../examples/tutorials/lighting/app';
-import ShaderHooksApp from '../../examples/tutorials/shader-hooks/app';
-import ShaderPluginsApp from '../../examples/tutorials/shader-plugins/app';
-import ShaderModulesApp from '../../examples/tutorials/shader-modules/app';
-import TransformFeedbackApp from '../../examples/tutorials/transform-feedback/app';
-import TransformApp from '../../examples/tutorials/transform/app';
-import {createArrowPathLayerDeck} from '../../examples/deck/arrow-path-layer/app';
-import {createArrowPolygonLayerDeck} from '../../examples/deck/arrow-polygon-layer/app';
-import {createArrowTextLayerDeck} from '../../examples/deck/arrow-text-layer/app';
-import {createGPUCulledTraceDeck} from '../../examples/deck/gpu-culled-trace/app';
+import {getErrorMessage, logError} from './react-luma/utils/error-utils';
 
 const exampleConfig = {};
+
+const loadAnimationApp = () => import('../../examples/api/animation/app');
+const loadBlendingApp = () => import('../../examples/api/blending/app');
+const loadCubemapApp = () => import('../../examples/api/cubemap/app');
+const loadArrowDggsPolygonsApp = () => import('../../examples/arrow/arrow-dggs-polygons/app');
+const loadArrowColumnRendererApp = () => import('../../examples/arrow/arrow-columns/app');
+const loadArrowMeshGeometryApp = () => import('../../examples/arrow/arrow-mesh-geometry/app');
+const loadArrowGeoArrowApp = () => import('../../examples/arrow/arrow-geoarrow/app');
+const loadArrowLinesApp = () => import('../../examples/arrow/arrow-lines/app');
+const loadArrowFloat64PrecisionApp = () => import('../../examples/arrow/arrow-float64-precision/app');
+const loadArrowPointRendererApp = () => import('../../examples/arrow/arrow-points/app');
+const loadArrowFilteringApp = () => import('../../examples/arrow/arrow-filtering/app');
+const loadArrowPolygonRendererApp = () => import('../../examples/arrow/arrow-polygons/app');
+const loadBloomApp = () => import('../../examples/experimental/bloom/app');
+const loadHTMLUIPrismApp = () => import('../../examples/experimental/html-ui-prism/app');
+const loadGPUFrustumCullingApp = () => import('../../examples/experimental/gpu-frustum-culling/app');
+const loadGPUSceneGraphApp = () => import('../../examples/experimental/gpu-scene-graph/app');
+const loadGPUTraceSceneApp = () => import('../../examples/experimental/gpu-trace-scene/app');
+const loadGPUTraceViewerApp = () => import('../../examples/experimental/gpu-trace-viewer/app');
+const loadGPUGraphExplorerApp = () => import('../../examples/experimental/gpu-graph-explorer/app');
+const loadLuCIMVolumeLabApp = () => import('../../examples/experimental/lucim-volume-lab/app');
+const loadGPT2App = () => import('../../examples/experimental/gpt-2/app');
+const loadVideoTextureApp = () => import('../../examples/api/video-texture/app');
+const loadWebXRKaleidoscopeApp = () => import('../../examples/experimental/webxr-kaleidoscope/app');
+const loadArrowParticlesApp = () => import('../../examples/arrow/arrow-particles/app');
+const loadMultiCanvasApp = () => import('../../examples/api/multi-canvas/app');
+const loadTexture3DApp = () => import('../../examples/api/texture-3d/app');
+const loadTextureSamplingApp = () => import('../../examples/api/texture-sampling/app');
+const loadTextureTesterApp = () => import('../../examples/api/texture-tester/app');
+const loadHelloReactApp = () => import('../../examples/integrations/hello-react/app');
+const loadDOFApp = () => import('../../examples/showcase/dof/app');
+const loadAdvancedEffectsApp = () => import('../../examples/experimental/advanced-effects/app');
+const loadDeferredRenderingApp = () => import('../../examples/experimental/deferred-rendering/app');
+const loadFluidFoundryApp = () => import('../../examples/experimental/fluid-foundry/app');
+const loadSpectralCausticsApp = () => import('../../examples/experimental/spectral-caustics/app');
+const loadVolumetricFireForgeApp = () => import('../../examples/experimental/volumetric-fire-forge/app');
+const loadVirtualGeometryCanyonApp = () => import('../../examples/experimental/virtual-geometry-canyon/app');
+const loadShadowMapApp = () => import('../../examples/experimental/shadow-map/app');
+const loadABufferApp = () => import('../../examples/experimental/a-buffer/app');
+const loadGLTFApp = () => import('../../examples/showcase/gltf/app');
+const loadGaussianSplatsApp = () => import('../../examples/showcase/gaussian-splats/app');
+const loadArrowTemporalStarfieldApp = () => import('../../examples/arrow/arrow-temporal-starfield/app');
+const loadArrowTimeColumnsApp = () => import('../../examples/arrow/arrow-time-columns/app');
+const loadArrowText2DApp = () => import('../../examples/arrow/arrow-text-2d/app');
+const loadInstancingApp = () => import('../../examples/showcase/instancing/app');
+const loadLightstormMegacityApp = () => import('../../examples/showcase/lightstorm-megacity/app');
+const loadLLMNetworkApp = () => import('../../examples/showcase/llm-network/app');
+const loadVectorFieldLabApp = () => import('../../examples/showcase/vector-field-lab/app');
+const loadQuantumStateStudioApp = () => import('../../examples/showcase/quantum-state-studio/app');
+const loadTempestOceanApp = () => import('../../examples/showcase/tempest-ocean/app');
+const loadRenderBundlesApp = () => import('../../examples/api/render-bundles/app');
+const loadTextSpaceCrawlApp = () => import('../../examples/experimental/text-space-crawl/app');
+const loadPersistenceApp = () => import('../../examples/showcase/persistence/app');
+const loadPostprocessingApp = () => import('../../examples/showcase/postprocessing/app');
+const loadAntialiasingApp = () => import('../../examples/experimental/antialiasing/app');
+const loadGlobeApp = () => import('../../examples/showcase/globe/app');
+const loadPacketSprayingApp = () => import('../../examples/showcase/packet-spraying/app');
+const loadHelloTriangleGeometryApp = () => import('../../examples/tutorials/hello-triangle-geometry/app');
+const loadHelloTriangleApp = () => import('../../examples/tutorials/hello-triangle/app');
+const loadHelloCubeApp = () => import('../../examples/tutorials/hello-cube/app');
+const loadTwoCubesApp = () => import('../../examples/tutorials/hello-two-cubes/app');
+const loadInstancedCubesApp = () => import('../../examples/tutorials/hello-instanced-cubes/app');
+const loadHelloInstancingApp = () => import('../../examples/tutorials/hello-instancing/app');
+const loadHelloGLTFApp = () => import('../../examples/tutorials/hello-gltf/app');
+const loadLightingApp = () => import('../../examples/tutorials/lighting/app');
+const loadShaderHooksApp = () => import('../../examples/tutorials/shader-hooks/app');
+const loadShaderPluginsApp = () => import('../../examples/tutorials/shader-plugins/app');
+const loadShaderModulesApp = () => import('../../examples/tutorials/shader-modules/app');
+const loadTransformFeedbackApp = () => import('../../examples/tutorials/transform-feedback/app');
+const loadTransformApp = () => import('../../examples/tutorials/transform/app');
 
 const loadBillionPointSpatialAtlasExample = () =>
   import('../../examples/showcase/billion-point-spatial-atlas/app');
@@ -127,6 +105,19 @@ const loadRasterLabExample = () => import('../../examples/showcase/raster-lab/ap
 const loadGPUSpatialTaxiExample = () => import('../../examples/deck/luspatial-taxi/app');
 const loadGPUGraphExplorerDeckExample = () => import('../../examples/deck/gpu-graph-explorer/app');
 const loadFP64Example = () => import('../../examples/experimental/fp64/app');
+const loadGPUSortExample = () => import('../../examples/experimental/gpu-sort/src/app');
+const loadGPUDataAnalysisExample = () =>
+  import('../../examples/experimental/gpu-data-analysis/src/app');
+const loadGPGPUShowcaseExample = () => import('../../examples/v10/gpgpu/src/app');
+const loadExternalContextExample = () =>
+  import('../../examples/integrations/external-context/app');
+const loadArrowPathLayerDeckExample = () => import('../../examples/deck/arrow-path-layer/app');
+const loadArrowPolygonLayerDeckExample = () =>
+  import('../../examples/deck/arrow-polygon-layer/app');
+const loadArrowTextLayerDeckExample = () => import('../../examples/deck/arrow-text-layer/app');
+const loadGPUCulledTraceDeckExample = () => import('../../examples/deck/gpu-culled-trace/app');
+
+const subscribeToInactiveApplication = (_listener: () => void): (() => void) => () => {};
 
 type WebsiteExampleProps = React.PropsWithChildren<
   ExampleDisplayProps & {
@@ -247,13 +238,32 @@ type CreateDeckExample = (
 ) =>
   | DeckExampleHandle
   | Promise<DeckExampleHandle>;
+
+const createDeferredArrowPathLayerDeck: CreateDeckExample = (parent, options) =>
+  loadArrowPathLayerDeckExample().then(({createArrowPathLayerDeck}) =>
+    createArrowPathLayerDeck(parent, options)
+  );
+const createDeferredArrowPolygonLayerDeck: CreateDeckExample = (parent, options) =>
+  loadArrowPolygonLayerDeckExample().then(({createArrowPolygonLayerDeck}) =>
+    createArrowPolygonLayerDeck(parent, options)
+  );
+const createDeferredArrowTextLayerDeck: CreateDeckExample = (parent, options) =>
+  loadArrowTextLayerDeckExample().then(({createArrowTextLayerDeck}) =>
+    createArrowTextLayerDeck(parent, options)
+  );
+const createDeferredGPUCulledTraceDeck: CreateDeckExample = (parent, options) =>
+  loadGPUCulledTraceDeckExample().then(({createGPUCulledTraceDeck}) =>
+    createGPUCulledTraceDeck(parent, options)
+  );
+
 type DeckArrowLayerPanelProps = {
   id: string;
   title: string;
   devices?: Array<'webgpu' | 'webgl2'>;
 };
 
-function makeDeckArrowLayerInfoPanel({id, title}: DeckArrowLayerPanelProps) {
+async function makeDeckArrowLayerInfoPanel({id, title}: DeckArrowLayerPanelProps) {
+  const {makeArrowExamplePanelHostHtml} = await import('../../examples/arrow/arrow-example-panels');
   return makeHtmlCustomPanel({
     id: `${id}-info`,
     title,
@@ -262,10 +272,26 @@ function makeDeckArrowLayerInfoPanel({id, title}: DeckArrowLayerPanelProps) {
 }
 
 function DeckArrowLayerPanel({id, title, devices = ['webgpu', 'webgl2']}: DeckArrowLayerPanelProps) {
-  const panel = useMemo(
-    () => makeDeckArrowLayerInfoPanel({id, title}),
-    [id, title]
-  );
+  const [panel, setPanel] = useState<Awaited<ReturnType<typeof makeDeckArrowLayerInfoPanel>>>();
+
+  useEffect(() => {
+    let isCancelled = false;
+    void makeDeckArrowLayerInfoPanel({id, title})
+      .then(nextPanel => {
+        if (!isCancelled) {
+          setPanel(nextPanel);
+        }
+      })
+      .catch(error => {
+        if (!isCancelled) {
+          logError(`Failed to load ${id} Arrow inspection panel`, error);
+        }
+      });
+
+    return () => {
+      isCancelled = true;
+    };
+  }, [id, title]);
 
   return (
     <div
@@ -335,13 +361,19 @@ function DeckArrowLayerCanvas({
 
     let isFinalized = false;
     let deck: DeckExampleHandle | null = null;
-    void Promise.resolve(createDeck(container, {device})).then(createdDeck => {
-      if (isFinalized) {
-        createdDeck.finalize();
-        return;
-      }
-      deck = createdDeck;
-    });
+    void Promise.resolve(createDeck(container, {device}))
+      .then(createdDeck => {
+        if (isFinalized) {
+          createdDeck.finalize();
+          return;
+        }
+        deck = createdDeck;
+      })
+      .catch(error => {
+        if (!isFinalized) {
+          logError(`Failed to initialize ${panel.id} Deck example`, error);
+        }
+      });
 
     return () => {
       isFinalized = true;
@@ -349,7 +381,7 @@ function DeckArrowLayerCanvas({
       container.replaceChildren();
       getCanvasContainer().appendChild(deviceCanvas);
     };
-  }, [createDeck, device]);
+  }, [createDeck, device, panel.id]);
 
   return (
     <>
@@ -379,7 +411,7 @@ export const DeckArrowPathLayerExample: React.FC<DeckArrowLayerExampleProps> = (
   <ReactExample
     component={DeckArrowLayerCanvas}
     componentProps={{
-      createDeck: createArrowPathLayerDeck,
+      createDeck: createDeferredArrowPathLayerDeck,
       panel: {
         id: 'arrow-path-layer',
         title: 'Arrow Path Layer'
@@ -396,7 +428,7 @@ export const DeckArrowPolygonLayerExample: React.FC<DeckArrowLayerExampleProps> 
   <ReactExample
     component={DeckArrowLayerCanvas}
     componentProps={{
-      createDeck: createArrowPolygonLayerDeck,
+      createDeck: createDeferredArrowPolygonLayerDeck,
       panel: {
         id: 'arrow-polygon-layer',
         title: 'Arrow Polygon Layer'
@@ -413,7 +445,7 @@ export const DeckArrowTextLayerExample: React.FC<DeckArrowLayerExampleProps> = (
   <ReactExample
     component={DeckArrowLayerCanvas}
     componentProps={{
-      createDeck: createArrowTextLayerDeck,
+      createDeck: createDeferredArrowTextLayerDeck,
       panel: {
         id: 'arrow-text-layer',
         title: 'Arrow Text Layer'
@@ -430,7 +462,7 @@ export const DeckGPUCulledTraceExample: React.FC<DeckArrowLayerExampleProps> = (
   <ReactExample
     component={DeckArrowLayerCanvas}
     componentProps={{
-      createDeck: createGPUCulledTraceDeck,
+      createDeck: createDeferredGPUCulledTraceDeck,
       panel: {
         id: 'gpu-culled-trace',
         title: 'GPU-Culled Trace with Arrow Text',
@@ -783,7 +815,7 @@ export const GLTFExample: React.FC<WebsiteExampleProps> = props => {
       title="glTF Asset Studio"
       subtitle="Physical materials · animated characters · standards-native glTF"
       directory="showcase"
-      template={GLTFApp}
+      loadTemplate={loadGLTFApp}
       config={exampleConfig}
       canvasContextProfile="high-dynamic-range"
       devices={referenceDevice ? [referenceDevice] : undefined}
@@ -825,7 +857,7 @@ export const GaussianSplatsExample: React.FC<WebsiteExampleProps> = props => {
       subtitle="Progressive HDR Gaussian splat rendering"
       directory="showcase"
       devices={['webgpu', 'webgl2']}
-      template={GaussianSplatsApp}
+      loadTemplate={loadGaussianSplatsApp}
       config={exampleConfig}
       canvasContextProfile="high-dynamic-range"
       showStats
@@ -838,17 +870,24 @@ export const GaussianSplatViewerExample: React.FC<
   WebsiteExampleProps & {defaultScene?: GaussianSplatSourceCatalogEntry['id']}
 > = ({defaultScene, ...props}) => {
   const loaderBundleUrl = useBaseUrl('/standalone-examples/gaussian-splats/loaders-gl.mjs');
-  const animationTemplate = useMemo(() => {
-    if (!defaultScene) {
-      return GaussianSplatsApp;
-    }
+  const loadAnimationTemplate = useMemo(
+    () => async () => {
+      const {default: GaussianSplatsApp} = await loadGaussianSplatsApp();
+      const makeAnimationTemplate = () => {
+        if (!defaultScene) {
+          return GaussianSplatsApp;
+        }
 
-    return class GaussianSplatSceneAnimationTemplate extends GaussianSplatsApp {
-      constructor(animationProps: AnimationProps) {
-        super({...animationProps, defaultScene});
-      }
-    };
-  }, [defaultScene]);
+        return class GaussianSplatSceneAnimationTemplate extends GaussianSplatsApp {
+          constructor(animationProps: AnimationProps) {
+            super({...animationProps, defaultScene});
+          }
+        };
+      };
+      return {default: makeAnimationTemplate()};
+    },
+    [defaultScene]
+  );
 
   if (typeof window !== 'undefined') {
     window.__lumaGaussianSplatsLoaderBundleUrl = loaderBundleUrl;
@@ -862,7 +901,7 @@ export const GaussianSplatViewerExample: React.FC<
       directory="showcase"
       sourcePath="examples/showcase/gaussian-splats/app.ts"
       devices={['webgpu', 'webgl2']}
-      template={animationTemplate}
+      loadTemplate={loadAnimationTemplate}
       config={exampleConfig}
       canvasContextProfile="high-dynamic-range"
       showStats
@@ -875,7 +914,7 @@ export const InstancingExample: React.FC = props => (
   <LumaExample
     id="instancing"
     directory="showcase"
-    template={InstancingApp}
+    loadTemplate={loadInstancingApp}
     config={exampleConfig}
     canvasContextProfile="high-dynamic-range"
     {...props}
@@ -889,7 +928,7 @@ export const LightstormMegacityExample: React.FC<WebsiteExampleProps> = props =>
     subtitle="GPU-driven city at data scale"
     directory="showcase"
     devices={['webgpu']}
-    template={LightstormMegacityApp}
+    loadTemplate={loadLightstormMegacityApp}
     config={exampleConfig}
     canvasContextProfile="high-dynamic-range"
     {...props}
@@ -903,7 +942,7 @@ export const VectorFieldLabExample: React.FC<WebsiteExampleProps> = props => (
     subtitle="Orbit linked 3D gradient, divergence, curl, and Laplacian volumes"
     directory="showcase"
     devices={['webgpu']}
-    template={VectorFieldLabApp}
+    loadTemplate={loadVectorFieldLabApp}
     config={exampleConfig}
     {...props}
   />
@@ -916,7 +955,7 @@ export const LLMNetworkExample: React.FC<WebsiteExampleProps> = props => (
     subtitle="Follow tokens through attention, hidden layers, and next-token prediction"
     directory="showcase"
     devices={['webgpu']}
-    template={LLMNetworkApp}
+    loadTemplate={loadLLMNetworkApp}
     config={exampleConfig}
     {...props}
   />
@@ -929,7 +968,7 @@ export const QuantumStateStudioExample: React.FC<WebsiteExampleProps> = props =>
     subtitle="GPU-resident state vectors · linked probability, phase, Bloch, and correlation views"
     directory="showcase"
     devices={['webgpu']}
-    template={QuantumStateStudioApp}
+    loadTemplate={loadQuantumStateStudioApp}
     config={exampleConfig}
     canvasContextProfile="high-dynamic-range"
     {...props}
@@ -1028,7 +1067,7 @@ export const TempestOceanExample: React.FC<WebsiteExampleProps> = props => (
     subtitle="GPUFFT2D displacement · HDR whitecaps"
     directory="showcase"
     devices={['webgpu']}
-    template={TempestOceanApp}
+    loadTemplate={loadTempestOceanApp}
     config={exampleConfig}
     canvasContextProfile="high-dynamic-range"
     {...props}
@@ -1040,7 +1079,7 @@ export const ArrowText2DExample: React.FC = props => (
     id="arrow-text-2d"
     title="Text"
     directory="arrow"
-    template={ArrowText2DApp}
+    loadTemplate={loadArrowText2DApp}
     config={exampleConfig}
     showStats
     {...props}
@@ -1052,7 +1091,7 @@ export const ArrowTimeColumnsExample: React.FC = props => (
     id="arrow-time-columns"
     title="Time"
     directory="arrow"
-    template={ArrowTimeColumnsApp}
+    loadTemplate={loadArrowTimeColumnsApp}
     config={exampleConfig}
     showStats
     {...props}
@@ -1064,7 +1103,7 @@ export const ArrowTemporalStarfieldExample: React.FC = props => (
     id="arrow-temporal-starfield"
     title="Durations"
     directory="arrow"
-    template={ArrowTemporalStarfieldApp}
+    loadTemplate={loadArrowTemporalStarfieldApp}
     config={exampleConfig}
     showStats
     {...props}
@@ -1076,7 +1115,7 @@ export const ArrowLinesExample: React.FC = props => (
     id="arrow-lines"
     title="Lines"
     directory="arrow"
-    template={ArrowLinesApp}
+    loadTemplate={loadArrowLinesApp}
     config={exampleConfig}
     showStats
     {...props}
@@ -1088,7 +1127,7 @@ export const ArrowFloat64PrecisionExample: React.FC = props => (
     id="arrow-float64-precision"
     title="Float64 Origin Rebasing"
     directory="arrow"
-    template={ArrowFloat64PrecisionApp}
+    loadTemplate={loadArrowFloat64PrecisionApp}
     config={exampleConfig}
     showStats
     {...props}
@@ -1100,7 +1139,7 @@ export const ArrowGeoArrowExample: React.FC<WebsiteExampleProps> = props => (
     id="arrow-geoarrow"
     title="GeoArrow"
     directory="arrow"
-    template={ArrowGeoArrowApp}
+    loadTemplate={loadArrowGeoArrowApp}
     config={exampleConfig}
     showStats
     {...props}
@@ -1112,7 +1151,7 @@ export const ArrowPointRendererExample: React.FC = props => (
     id="arrow-points"
     title="Points"
     directory="arrow"
-    template={ArrowPointRendererApp}
+    loadTemplate={loadArrowPointRendererApp}
     config={exampleConfig}
     showStats
     {...props}
@@ -1124,7 +1163,7 @@ export const ArrowFilteringExample: React.FC = props => (
     id="arrow-filtering"
     title="ShaderPlugin Filtering"
     directory="arrow"
-    template={ArrowFilteringApp}
+    loadTemplate={loadArrowFilteringApp}
     config={exampleConfig}
     showStats
     {...props}
@@ -1136,7 +1175,7 @@ export const ArrowColumnRendererExample: React.FC = props => (
     id="arrow-columns"
     title="DGGS + time"
     directory="arrow"
-    template={ArrowColumnRendererApp}
+    loadTemplate={loadArrowColumnRendererApp}
     config={exampleConfig}
     devices={['webgpu']}
     showStats
@@ -1149,7 +1188,7 @@ export const ArrowPolygonRendererExample: React.FC = props => (
     id="arrow-polygons"
     title="Polygons"
     directory="arrow"
-    template={ArrowPolygonRendererApp}
+    loadTemplate={loadArrowPolygonRendererApp}
     config={exampleConfig}
     showStats
     {...props}
@@ -1161,7 +1200,7 @@ export const ArrowDggsPolygonsExample: React.FC = props => (
     id="arrow-dggs-polygons"
     title="Global Grids"
     directory="arrow"
-    template={ArrowDggsPolygonsApp}
+    loadTemplate={loadArrowDggsPolygonsApp}
     config={exampleConfig}
     devices={['webgpu']}
     showStats
@@ -1179,16 +1218,24 @@ export const GPGPUExample: React.FC = () => {
       return;
     }
 
+    let isCancelled = false;
     let handle: GPGPUShowcaseHandle | null = null;
-    try {
-      setErrorMessage(null);
-      handle = initializeGPGPUShowcase({device});
-    } catch (error) {
-      setErrorMessage(getErrorMessage(error));
-      logError('Failed to initialize GPGPU example', error);
-    }
+    setErrorMessage(null);
+    void loadGPGPUShowcaseExample()
+      .then(({initializeGPGPUShowcase}) => {
+        if (!isCancelled) {
+          handle = initializeGPGPUShowcase({device});
+        }
+      })
+      .catch(error => {
+        if (!isCancelled) {
+          setErrorMessage(getErrorMessage(error));
+          logError('Failed to initialize GPGPU example', error);
+        }
+      });
 
     return () => {
+      isCancelled = true;
       handle?.destroy();
     };
   }, [deviceType, device]);
@@ -1267,15 +1314,23 @@ export const GPUSortExample: React.FC<WebsiteExampleProps> = ({embeddedHeight, .
   const [errorMessage, setErrorMessage] = useState<string | null>(null);
 
   useEffect(() => {
+    let isCancelled = false;
     let handle: GPUSortExampleHandle | null = null;
-    try {
-      handle = initializeGPUSortExample();
-    } catch (error) {
-      setErrorMessage(getErrorMessage(error));
-      logError('Failed to initialize GPU sort example', error);
-    }
+    void loadGPUSortExample()
+      .then(({initializeGPUSortExample}) => {
+        if (!isCancelled) {
+          handle = initializeGPUSortExample();
+        }
+      })
+      .catch(error => {
+        if (!isCancelled) {
+          setErrorMessage(getErrorMessage(error));
+          logError('Failed to initialize GPU sort example', error);
+        }
+      });
 
     return () => {
+      isCancelled = true;
       handle?.destroy();
     };
   }, []);
@@ -1304,14 +1359,24 @@ export const GPUDataAnalysisExample: React.FC<WebsiteExampleProps> = ({
   const [errorMessage, setErrorMessage] = useState<string | null>(null);
 
   useEffect(() => {
+    let isCancelled = false;
     let handle: GPUDataAnalysisExampleHandle | null = null;
-    try {
-      handle = initializeGPUDataAnalysisExample();
-    } catch (error) {
-      setErrorMessage(getErrorMessage(error));
-      logError('Failed to initialize GPU data-analysis example', error);
-    }
-    return () => handle?.destroy();
+    void loadGPUDataAnalysisExample()
+      .then(({initializeGPUDataAnalysisExample}) => {
+        if (!isCancelled) {
+          handle = initializeGPUDataAnalysisExample();
+        }
+      })
+      .catch(error => {
+        if (!isCancelled) {
+          setErrorMessage(getErrorMessage(error));
+          logError('Failed to initialize GPU data-analysis example', error);
+        }
+      });
+    return () => {
+      isCancelled = true;
+      handle?.destroy();
+    };
   }, []);
 
   return (
@@ -1339,7 +1404,7 @@ export const GPT2Example: React.FC = props => (
     showHeader={false}
     showStats={false}
     templateInfoPlacement="page"
-    template={GPT2App}
+    loadTemplate={loadGPT2App}
     config={exampleConfig}
     {...props}
   />
@@ -1350,7 +1415,7 @@ export const TextSpaceCrawlExample: React.FC = props => (
     id="text-space-crawl"
     title="Text Space Crawl"
     directory="experimental"
-    template={TextSpaceCrawlApp}
+    loadTemplate={loadTextSpaceCrawlApp}
     config={exampleConfig}
     {...props}
   />
@@ -1360,7 +1425,7 @@ export const PersistenceExample: React.FC<WebsiteExampleProps> = props => (
   <LumaExample
     id="persistence"
     directory="showcase"
-    template={PersistenceApp}
+    loadTemplate={loadPersistenceApp}
     config={exampleConfig}
     {...props}
   />
@@ -1370,22 +1435,28 @@ export const PostprocessingExample: React.FC<WebsiteExampleProps & {effect?: str
   effect,
   ...props
 }) => {
-  const template = useMemo(() => {
-    if (!effect) {
-      return PostprocessingApp;
-    }
+  const loadTemplate = useMemo(
+    () => async () => {
+      const {default: PostprocessingApp} = await loadPostprocessingApp();
+      if (!effect) {
+        return {default: PostprocessingApp};
+      }
 
-    return class FocusedPostprocessingApp extends PostprocessingApp {
-      static override initialEffectName = effect;
-    };
-  }, [effect]);
+      return {
+        default: class FocusedPostprocessingApp extends PostprocessingApp {
+          static override initialEffectName = effect;
+        }
+      };
+    },
+    [effect]
+  );
 
   return (
     <LumaExample
       id="postprocessing"
       title="Effects: Image Processing"
       directory="showcase"
-      template={template}
+      loadTemplate={loadTemplate}
       config={exampleConfig}
       {...props}
     />
@@ -1397,7 +1468,7 @@ export const AntialiasingExample: React.FC<WebsiteExampleProps> = props => (
     id="antialiasing"
     title="Antialiasing Techniques"
     directory="experimental"
-    template={AntialiasingApp}
+    loadTemplate={loadAntialiasingApp}
     config={exampleConfig}
     {...props}
   />
@@ -1408,7 +1479,7 @@ export const GlobeExample: React.FC = props => (
     id="globe"
     title="Globe"
     directory="showcase"
-    template={GlobeApp}
+    loadTemplate={loadGlobeApp}
     config={exampleConfig}
     canvasContextProfile="high-dynamic-range"
     {...props}
@@ -1421,7 +1492,7 @@ export const PacketSprayingExample: React.FC<WebsiteExampleProps> = props => (
     title="Effects: Glass"
     subtitle="Network Packet Spraying"
     directory="showcase"
-    template={PacketSprayingApp}
+    loadTemplate={loadPacketSprayingApp}
     config={exampleConfig}
     canvasContextProfile="high-dynamic-range"
     {...props}
@@ -1433,7 +1504,7 @@ export const DOFExample: React.FC<WebsiteExampleProps> = props => (
     id="dof"
     title="Depth of Field"
     directory="showcase"
-    template={DOFApp}
+    loadTemplate={loadDOFApp}
     config={exampleConfig}
     {...props}
   />
@@ -1444,7 +1515,7 @@ export const AdvancedEffectsExample: React.FC<WebsiteExampleProps> = props => (
     id="advanced-effects"
     title="Advanced Effects: Visualization City"
     directory="experimental"
-    template={AdvancedEffectsApp}
+    loadTemplate={loadAdvancedEffectsApp}
     config={exampleConfig}
     devices={['webgpu-max']}
     requiredDeviceLimits={{maxColorAttachmentBytesPerSample: 44}}
@@ -1457,7 +1528,7 @@ export const DeferredRenderingExample: React.FC<WebsiteExampleProps> = props => 
     id="deferred-rendering"
     title="Deferred Rendering: Illumination Lab"
     directory="experimental"
-    template={DeferredRenderingApp}
+    loadTemplate={loadDeferredRenderingApp}
     config={exampleConfig}
     devices={['webgpu-max', 'webgpu-core']}
     requiredDeviceLimits={{
@@ -1475,7 +1546,7 @@ export const FluidFoundryExample: React.FC<WebsiteExampleProps> = props => (
     title="Fluid Foundry: Liquid Metal Press"
     subtitle="GPU-resident MLS-MPM fluid"
     directory="experimental"
-    template={FluidFoundryApp}
+    loadTemplate={loadFluidFoundryApp}
     config={exampleConfig}
     devices={['webgpu']}
     canvasContextProfile="high-dynamic-range"
@@ -1488,7 +1559,7 @@ export const SpectralCausticsExample: React.FC<WebsiteExampleProps> = props => (
     id="spectral-caustics"
     title="Spectral Caustics: Prism Cathedral"
     directory="experimental"
-    template={SpectralCausticsApp}
+    loadTemplate={loadSpectralCausticsApp}
     config={exampleConfig}
     devices={['webgpu']}
     canvasContextProfile="high-dynamic-range"
@@ -1502,7 +1573,7 @@ export const VolumetricFireForgeExample: React.FC<WebsiteExampleProps> = props =
     title="Volumetric Fire Forge"
     subtitle="Reactive HDR fire on the GPU"
     directory="experimental"
-    template={VolumetricFireForgeApp}
+    loadTemplate={loadVolumetricFireForgeApp}
     config={exampleConfig}
     devices={['webgpu']}
     canvasContextProfile="high-dynamic-range"
@@ -1516,7 +1587,7 @@ export const LuCIMVolumeLabExample: React.FC<WebsiteExampleProps> = props => (
     title="LuCIM Volume Lab"
     subtitle="GPU-resident tri-planar volume segmentation"
     directory="experimental"
-    template={LuCIMVolumeLabApp}
+    loadTemplate={loadLuCIMVolumeLabApp}
     config={exampleConfig}
     devices={['webgpu']}
     {...props}
@@ -1529,7 +1600,7 @@ export const VirtualGeometryCanyonExample: React.FC<WebsiteExampleProps> = props
     title="Virtual Geometry Canyon"
     subtitle="GPU-driven hierarchical terrain LOD"
     directory="experimental"
-    template={VirtualGeometryCanyonApp}
+    loadTemplate={loadVirtualGeometryCanyonApp}
     config={exampleConfig}
     devices={['webgpu']}
     {...props}
@@ -1541,7 +1612,7 @@ export const ShadowMapExample: React.FC<WebsiteExampleProps> = props => (
     id="shadow-map"
     title="Effects: Shadow Map Quality"
     directory="experimental"
-    template={ShadowMapApp}
+    loadTemplate={loadShadowMapApp}
     config={exampleConfig}
     devices={['webgpu']}
     {...props}
@@ -1553,7 +1624,7 @@ export const OITExample: React.FC<WebsiteExampleProps> = props => (
     id="a-buffer"
     title="Order-independent Transparency"
     directory="experimental"
-    template={ABufferApp}
+    loadTemplate={loadABufferApp}
     config={exampleConfig}
     devices={['webgpu', 'webgl2']}
     {...props}
@@ -1565,7 +1636,7 @@ export const BloomExample: React.FC<WebsiteExampleProps> = props => (
     id="bloom"
     title="Bloom"
     directory="experimental"
-    template={BloomApp}
+    loadTemplate={loadBloomApp}
     config={exampleConfig}
     canvasContextProfile="high-dynamic-range"
     {...props}
@@ -1577,7 +1648,7 @@ export const VideoTextureExample: React.FC<WebsiteExampleProps> = props => {
   const [cameraError, setCameraError] = useState<string | null>(null);
   const [isCameraBlocked, setIsCameraBlocked] = useState(false);
   const handleUseCamera = async () => {
-    const app = VideoTextureApp.current;
+    const app = (await loadVideoTextureApp()).default.current;
     if (!app) {
       setCameraStatus('error');
       setCameraError('Example is still starting');
@@ -1602,7 +1673,7 @@ export const VideoTextureExample: React.FC<WebsiteExampleProps> = props => {
       id="video-texture"
       title="Video Texture"
       directory="api"
-      template={VideoTextureApp}
+      loadTemplate={loadVideoTextureApp}
       config={exampleConfig}
       headerControls={
         <div style={{display: 'flex', alignItems: 'center', gap: 10, marginTop: 12}}>
@@ -1646,10 +1717,12 @@ export const WebXRKaleidoscopeExample: React.FC = props => {
   type XRStatus = 'idle' | 'pending' | 'live' | 'error';
   type XRMode = 'immersive-ar' | 'immersive-vr';
 
+  const {module: webXRModule} = useDeferredExampleModule(loadWebXRKaleidoscopeApp);
+  const webXRApplication = webXRModule?.default;
   const selectedDevice = useStore(store => store.device);
   const activeApplication = useSyncExternalStore(
-    WebXRKaleidoscopeApp.subscribeToCurrent,
-    () => WebXRKaleidoscopeApp.current,
+    webXRApplication?.subscribeToCurrent || subscribeToInactiveApplication,
+    () => webXRApplication?.current || null,
     () => null
   );
   const [xrStatus, setXRStatus] = useState<XRStatus>('idle');
@@ -1671,7 +1744,7 @@ export const WebXRKaleidoscopeExample: React.FC = props => {
     : 'WebGL2 · immersive stereo and optional AR camera access';
 
   const handleXRSession = async (sessionMode: XRMode) => {
-    const app = WebXRKaleidoscopeApp.current;
+    const app = webXRApplication?.current;
     if (!app) {
       setXRStatus('error');
       setXRError('Example is still starting');
@@ -1735,7 +1808,7 @@ export const WebXRKaleidoscopeExample: React.FC = props => {
       directory="experimental"
       devices={['webgpu', 'webgl2']}
       xrCompatible
-      template={WebXRKaleidoscopeApp}
+      loadTemplate={loadWebXRKaleidoscopeApp}
       config={exampleConfig}
       headerControls={
         <div
@@ -1820,7 +1893,7 @@ export const HTMLUIPrismExample: React.FC<WebsiteExampleProps> = props => (
     id="html-ui-prism"
     title="HTML-in-Canvas Prism"
     directory="experimental"
-    template={HTMLUIPrismApp}
+    loadTemplate={loadHTMLUIPrismApp}
     config={exampleConfig}
     {...props}
   />
@@ -1832,7 +1905,7 @@ export const GPUTraceViewerExample: React.FC = props => (
     title="GPU Hierarchical Trace Viewer"
     directory="experimental"
     devices={['webgpu-max']}
-    template={GPUTraceViewerApp}
+    loadTemplate={loadGPUTraceViewerApp}
     config={exampleConfig}
     {...props}
   />
@@ -1845,7 +1918,7 @@ export const GPUGraphExplorerExample: React.FC<WebsiteExampleProps> = props => (
     subtitle="GPU-native topology, analytics, selection, and progressive force layout"
     directory="experimental"
     devices={['webgpu']}
-    template={GPUGraphExplorerApp}
+    loadTemplate={loadGPUGraphExplorerApp}
     config={exampleConfig}
     {...props}
   />
@@ -1857,7 +1930,7 @@ export const GPUTraceSceneExample: React.FC = props => (
     title="GPU Scene Trace Explorer"
     directory="experimental"
     devices={['webgpu-max']}
-    template={GPUTraceSceneApp}
+    loadTemplate={loadGPUTraceSceneApp}
     config={exampleConfig}
     {...props}
   />
@@ -1869,7 +1942,7 @@ export const GPUSceneGraphExample: React.FC = props => (
     title="GPU Scene Graph Explorer"
     directory="experimental"
     devices={['webgpu-max']}
-    template={GPUSceneGraphApp}
+    loadTemplate={loadGPUSceneGraphApp}
     config={exampleConfig}
     {...props}
   />
@@ -1881,7 +1954,7 @@ export const GPUFrustumCullingExample: React.FC = props => (
     title="GPU Frustum Culling"
     directory="experimental"
     devices={['webgpu']}
-    template={GPUFrustumCullingApp}
+    loadTemplate={loadGPUFrustumCullingApp}
     config={exampleConfig}
     {...props}
   />
@@ -1892,7 +1965,7 @@ export const ArrowMeshGeometryExample: React.FC = props => (
     id="arrow-mesh-geometry"
     title="Matrices"
     directory="arrow"
-    template={ArrowMeshGeometryApp}
+    loadTemplate={loadArrowMeshGeometryApp}
     config={exampleConfig}
     {...props}
   />
@@ -1903,7 +1976,7 @@ export const ArrowParticlesExample: React.FC = props => (
     id="arrow-particles"
     title="Particles"
     directory="arrow"
-    template={ArrowParticlesApp}
+    loadTemplate={loadArrowParticlesApp}
     config={exampleConfig}
     {...props}
   />
@@ -1915,7 +1988,7 @@ export const AnimationExample: React.FC<WebsiteExampleProps> = props => (
   <LumaExample
     id="animation"
     directory="api"
-    template={AnimationApp}
+    loadTemplate={loadAnimationApp}
     config={exampleConfig}
     showStats
     {...props}
@@ -1927,7 +2000,7 @@ export const BlendingExample: React.FC<WebsiteExampleProps> = props => (
     id="blending"
     title="Blending"
     directory="api"
-    template={BlendingApp}
+    loadTemplate={loadBlendingApp}
     config={exampleConfig}
     devices={['webgpu', 'webgl2']}
     {...props}
@@ -1939,13 +2012,15 @@ export const CubemapExample: React.FC<WebsiteExampleProps> = props => (
     id="cubemap"
     title="Texture Cube"
     directory="api"
-    template={CubemapApp}
+    loadTemplate={loadCubemapApp}
     config={exampleConfig}
     {...props}
   />
 );
 
 export const MultiCanvasExample: React.FC<WebsiteExampleProps> = props => {
+  const {module: multiCanvasModule, errorMessage} = useDeferredExampleModule(loadMultiCanvasApp);
+  const MultiCanvasApp = multiCanvasModule?.default;
   const deviceType = useStore(store => store.deviceType);
   const presentationDevice = useStore(store => store.presentationDevice);
   const presentationDeviceError = useStore(store => store.presentationDeviceError);
@@ -1956,15 +2031,15 @@ export const MultiCanvasExample: React.FC<WebsiteExampleProps> = props => {
     style: props.style
   };
 
-  if (presentationDeviceError) {
+  if (presentationDeviceError || errorMessage) {
     return (
       <ExamplePage {...exampleDisplayProps}>
-        <div>{presentationDeviceError}</div>
+        <div>{presentationDeviceError || errorMessage}</div>
       </ExamplePage>
     );
   }
 
-  return deviceType && presentationDevice ? (
+  return deviceType && presentationDevice && MultiCanvasApp ? (
     <ReactExample
       component={MultiCanvasApp}
       componentProps={{deviceType: getExampleDeviceType(presentationDevice), presentationDevice}}
@@ -2092,7 +2167,7 @@ export const Texture3DExample: React.FC<WebsiteExampleProps> = props => (
     id="texture-3d"
     directory="api-3d"
     sourceDirectory="api"
-    template={Texture3DApp}
+    loadTemplate={loadTexture3DApp}
     config={exampleConfig}
     {...props}
   />
@@ -2103,7 +2178,7 @@ export const TextureSamplingExample: React.FC<WebsiteExampleProps> = props => (
     id="texture-sampling"
     title="Texture Sampling"
     directory="api"
-    template={TextureSamplingApp}
+    loadTemplate={loadTextureSamplingApp}
     config={exampleConfig}
     devices={['webgpu', 'webgl2']}
     {...props}
@@ -2114,6 +2189,8 @@ export const TextureTesterExample: React.FC<WebsiteExampleProps> = ({
   embeddedHeight,
   ...props
 }) => {
+  const {module: textureTesterModule, errorMessage} = useDeferredExampleModule(loadTextureTesterApp);
+  const TextureTesterApp = textureTesterModule?.default;
   const deviceType = useStore(store => store.deviceType);
   const presentationDevice = useStore(store => store.presentationDevice);
   const presentationDeviceError = useStore(store => store.presentationDeviceError);
@@ -2146,9 +2223,9 @@ export const TextureTesterExample: React.FC<WebsiteExampleProps> = ({
           />
         </div>
       ) : null}
-      {presentationDeviceError ? (
-        <div>{presentationDeviceError}</div>
-      ) : deviceType && presentationDevice ? (
+      {presentationDeviceError || errorMessage ? (
+        <div>{presentationDeviceError || errorMessage}</div>
+      ) : deviceType && presentationDevice && TextureTesterApp ? (
         <TextureTesterApp
           compact={props.embedded}
           deviceType={getExampleDeviceType(presentationDevice)}
@@ -2166,7 +2243,7 @@ export const RenderBundlesExample: React.FC<WebsiteExampleProps> = props => (
     id="render-bundles"
     title="Render Bundles"
     directory="api"
-    template={RenderBundlesApp}
+    loadTemplate={loadRenderBundlesApp}
     config={exampleConfig}
     devices={['webgpu']}
     showStats
@@ -2184,18 +2261,29 @@ export const ExternalContextExample: React.FC = () => {
     const container = containerRef.current;
     if (!container) return undefined;
 
+    let isCancelled = false;
     let exampleHandle: ExternalWebGLContextHandle | null = null;
 
-    initializeExternalWebGLContext({container})
+    loadExternalContextExample()
+      .then(({default: initializeExternalWebGLContext}) =>
+        initializeExternalWebGLContext({container})
+      )
       .then(instance => {
+        if (isCancelled) {
+          instance.destroy();
+          return;
+        }
         exampleHandle = instance;
       })
       .catch(caughtError => {
-        logError('External WebGL context example failed', caughtError);
-        setError(getErrorMessage(caughtError));
+        if (!isCancelled) {
+          logError('External WebGL context example failed', caughtError);
+          setError(getErrorMessage(caughtError));
+        }
       });
 
     return () => {
+      isCancelled = true;
       exampleHandle?.destroy();
     };
   }, []);
@@ -2214,6 +2302,8 @@ export const ExternalContextExample: React.FC = () => {
 };
 
 export const ReactStrictModeExample: React.FC = () => {
+  const {module: reactExampleModule, errorMessage} = useDeferredExampleModule(loadHelloReactApp);
+  const HelloReactApp = reactExampleModule?.default;
   const [showCube, setShowCube] = useState(true);
   const [mountCount, setMountCount] = useState(0);
 
@@ -2262,14 +2352,20 @@ export const ReactStrictModeExample: React.FC = () => {
         </div>
       </ExampleHeader>
       <div className="integration-example-page" style={{width: '100%', minHeight: '640px'}}>
-        <React.StrictMode>
-          <HelloReactApp
-            showControls={false}
-            showCube={showCube}
-            mountCount={mountCount}
-            onToggleCube={toggleCube}
-          />
-        </React.StrictMode>
+        {errorMessage ? (
+          <div role="alert">{errorMessage}</div>
+        ) : HelloReactApp ? (
+          <React.StrictMode>
+            <HelloReactApp
+              showControls={false}
+              showCube={showCube}
+              mountCount={mountCount}
+              onToggleCube={toggleCube}
+            />
+          </React.StrictMode>
+        ) : (
+          <div>Loading React example...</div>
+        )}
       </div>
     </ExamplePage>
   );
@@ -2281,7 +2377,7 @@ export const HelloTriangleExample: React.FC = props => (
   <LumaExample
     id="hello-triangle"
     directory="tutorials"
-    template={HelloTriangleApp}
+    loadTemplate={loadHelloTriangleApp}
     config={exampleConfig}
     showStats={false}
     stackBlitz
@@ -2297,7 +2393,7 @@ export const HelloTriangleGeometryExample: React.FC = props => (
   <LumaExample
     id="hello-triangle-geometry"
     directory="tutorials"
-    template={HelloTriangleGeometryApp}
+    loadTemplate={loadHelloTriangleGeometryApp}
     config={exampleConfig}
     showStats={false}
     stackBlitz
@@ -2309,7 +2405,7 @@ export const HelloCubeExample: React.FC = props => (
   <LumaExample
     id="hello-cube"
     directory="tutorials"
-    template={HelloCubeApp}
+    loadTemplate={loadHelloCubeApp}
     config={exampleConfig}
     showStats={false}
     stackBlitz
@@ -2321,7 +2417,7 @@ export const InstancedCubesExample: React.FC = props => (
   <LumaExample
     id="instanced-cubes"
     directory="tutorials"
-    template={InstancedCubesApp}
+    loadTemplate={loadInstancedCubesApp}
     config={exampleConfig}
     showStats={false}
     stackBlitz
@@ -2333,7 +2429,7 @@ export const TwoCubesExample: React.FC = props => (
   <LumaExample
     id="two-cubes"
     directory="tutorials"
-    template={TwoCubesApp}
+    loadTemplate={loadTwoCubesApp}
     config={exampleConfig}
     showStats={false}
     stackBlitz
@@ -2345,7 +2441,7 @@ export const LightingExample: React.FC = props => (
   <LumaExample
     id="lighting"
     directory="tutorials"
-    template={LightingApp}
+    loadTemplate={loadLightingApp}
     config={exampleConfig}
     showStats={false}
     stackBlitz
@@ -2357,7 +2453,7 @@ export const HelloGLTFExample: React.FC = props => (
   <LumaExample
     id="hello-gltf"
     directory="tutorials"
-    template={HelloGLTFApp}
+    loadTemplate={loadHelloGLTFApp}
     config={exampleConfig}
     showStats={false}
     stackBlitz
@@ -2369,7 +2465,7 @@ export const HelloInstancingExample: React.FC = props => (
   <LumaExample
     id="hello-instancing"
     directory="tutorials"
-    template={HelloInstancingApp}
+    loadTemplate={loadHelloInstancingApp}
     config={exampleConfig}
     showStats={false}
     stackBlitz
@@ -2381,7 +2477,7 @@ export const ShaderHooksExample: React.FC = props => (
   <LumaExample
     id="shader-hooks"
     directory="tutorials"
-    template={ShaderHooksApp}
+    loadTemplate={loadShaderHooksApp}
     config={exampleConfig}
     showStats={false}
     stackBlitz
@@ -2393,7 +2489,7 @@ export const ShaderPluginsExample: React.FC = props => (
   <LumaExample
     id="shader-plugins"
     directory="tutorials"
-    template={ShaderPluginsApp}
+    loadTemplate={loadShaderPluginsApp}
     config={exampleConfig}
     showStats={false}
     stackBlitz
@@ -2405,7 +2501,7 @@ export const ShaderModulesExample: React.FC = props => (
   <LumaExample
     id="shader-modules"
     directory="tutorials"
-    template={ShaderModulesApp}
+    loadTemplate={loadShaderModulesApp}
     config={exampleConfig}
     showStats={false}
     stackBlitz
@@ -2417,7 +2513,7 @@ export const TransformFeedbackExample: React.FC = props => (
   <LumaExample
     id="transform-feedback"
     directory="tutorials"
-    template={TransformFeedbackApp}
+    loadTemplate={loadTransformFeedbackApp}
     config={exampleConfig}
     showStats={false}
     stackBlitz
@@ -2430,7 +2526,7 @@ export const TransformExample: React.FC = props => (
   <LumaExample
     id="transform"
     directory="tutorials"
-    template={TransformApp}
+    loadTemplate={loadTransformApp}
     config={exampleConfig}
     showStats={false}
     stackBlitz
