+++ website/src/react-luma/components/device-tabs.tsx
@@ -93,27 +93,49 @@ export const DeviceTabsPriv = (props: DeviceTabsProps = {}) => {
   return (
     <div
       ref={setDeviceTabsElement}
+      data-luma-device-selector=""
       data-luma-example-appearance={appearance}
       style={props.style}
     >
-      <Tabs
-        label="Graphics backend"
-        selectedItem={selectedDeviceType}
-        setSelectedItem={selectDeviceType}
-      >
-        {devices.map(type => (
-          <Tab
-            key={type}
-            title={DEVICE_TAB_LABELS[type]}
-            tag={type}
-            badge={DEVICE_TAB_BADGES[type]}
-            unavailableBadge={deviceAvailability[type] === false ? 'N/A' : undefined}
-            disabled={deviceAvailability[type] === false}
-          >
-            {type === deviceType ? deviceError : null}
-          </Tab>
-        ))}
-      </Tabs>
+      <div data-luma-device-tabs="">
+        <Tabs
+          label="Graphics backend"
+          selectedItem={selectedDeviceType}
+          setSelectedItem={selectDeviceType}
+        >
+          {devices.map(type => (
+            <Tab
+              key={type}
+              title={DEVICE_TAB_LABELS[type]}
+              tag={type}
+              badge={DEVICE_TAB_BADGES[type]}
+              unavailableBadge={deviceAvailability[type] === false ? 'N/A' : undefined}
+              disabled={deviceAvailability[type] === false}
+            >
+              {type === deviceType ? deviceError : null}
+            </Tab>
+          ))}
+        </Tabs>
+      </div>
+      <label data-luma-device-menu="">
+        <span data-luma-device-menu-indicator="" aria-hidden="true" />
+        <select
+          aria-label="Graphics backend"
+          data-luma-device-select=""
+          value={selectedDeviceType ?? devices[0] ?? ''}
+          onChange={event => void selectDeviceType(event.currentTarget.value as DeviceType)}
+        >
+          {devices.map(type => (
+            <option
+              key={type}
+              value={type}
+              disabled={deviceAvailability[type] === false}
+            >
+              {getDeviceOptionLabel(type)}
+            </option>
+          ))}
+        </select>
+      </label>
     </div>
   );
 };
@@ -144,4 +166,9 @@ function getDeviceTypes(devices?: DeviceTabsProps['devices']): DeviceType[] {
   }
 
   return deviceTypes;
+}
+
+function getDeviceOptionLabel(deviceType: DeviceType): string {
+  const badge = DEVICE_TAB_BADGES[deviceType];
+  return badge ? `${DEVICE_TAB_LABELS[deviceType]} ${badge}` : DEVICE_TAB_LABELS[deviceType];
 }
