+++ website/src/react-luma/components/info-box.tsx
@@ -417,6 +417,7 @@ function InfoBoxView(props: InfoBoxViewProps) {
     <div
       ref={setInfoBoxElement}
       data-info-box-appearance={appearance}
+      data-luma-info-box-collapsed={isCollapsed ? 'true' : 'false'}
       style={{
         ...INFO_BOX_BASE_STYLE,
         ...INFO_BOX_APPEARANCE_STYLES[appearance],
@@ -432,6 +433,7 @@ function InfoBoxView(props: InfoBoxViewProps) {
     >
       <style>{INFO_BOX_CHROME_STYLE}</style>
       <div
+        data-luma-example-info-header=""
         style={{
           display: 'flex',
           alignItems: 'flex-start',
@@ -470,6 +472,7 @@ function InfoBoxView(props: InfoBoxViewProps) {
                 }}
               />
               <h3
+                data-luma-example-title=""
                 style={{
                   color: 'inherit',
                   fontSize: 17,
@@ -486,6 +489,7 @@ function InfoBoxView(props: InfoBoxViewProps) {
           ) : null}
           {props.subtitle ? (
             <div
+              data-luma-example-subtitle=""
               style={{
                 color: `var(--luma-example-text-muted, ${theme.textMuted})`,
                 fontSize: 12,
@@ -498,10 +502,14 @@ function InfoBoxView(props: InfoBoxViewProps) {
             </div>
           ) : null}
         </button>
-        <div style={{display: 'flex', alignItems: 'center', gap: 8, flexShrink: 0}}>
+        <div
+          data-luma-example-chrome-actions=""
+          style={{display: 'flex', alignItems: 'center', gap: 8, flexShrink: 0}}
+        >
           {sourceUrl ? (
             <a
               data-luma-example-chrome-action=""
+              data-luma-example-source-link=""
               href={sourceUrl}
               target="_blank"
               rel="noreferrer"
@@ -526,11 +534,12 @@ function InfoBoxView(props: InfoBoxViewProps) {
                   fill="currentColor"
                 />
               </svg>
-              GitHub
+              <span data-luma-example-source-label="">GitHub</span>
             </a>
           ) : null}
           <button
             data-luma-example-chrome-action=""
+            data-luma-example-info-toggle=""
             type="button"
             aria-label={isCollapsed ? 'Expand info box' : 'Collapse info box'}
             aria-expanded={!isCollapsed}
